rand func in game sys update
change it to template

+++ Nikesaurus/src/Systems/GameLogic/sysGameLogic.cpp
@@ -98,8 +98,8 @@ namespace NIKE {
 			NIKE_METADATA_SERVICE->addEntityTag(enemy_entity, "enemy");
 		}
 		// Randomly offset from spawner position
-		float offset_x = static_cast<float>(getRandomNumber(-20, 20));
-		float offset_y = static_cast<float>(getRandomNumber(-20, 20));
+		float offset_x = getRandomNumber(-20.f, 20.f);
+		float offset_y = getRandomNumber(-20.f, 20.f);
 
 		// Set Enemy Position
 		auto enemy_transform_comp = NIKE_ECS_MANAGER->getEntityComponent<Transform::Transform>(enemy_entity);
